Discover palette

Discover palette

Usage

hermitage_palette("parsons_2")
hermitage_palette("hermitage_1")
hermitage_palette("harmony", 5)
hermitage_palette("imari", 5)
hermitage_palette("du_barry", 3)
For more colours:
more_colours <- hermitage_palette(25, name = "imari", type = "continuous")

Format

Arguments

name

Name of the palette. Options: parsons_1, parsons_2, du_barry, delft_ware, imari, collioure_matisse, peacock_clock, prodigal_son, harmony , dance_matisse , judith, magdalene_titian, bagration, flight_gauguin, hermitage_1, hermitage_2, madonna_litta, cottages_vincent, danae, battista_cima, weyden, faberge, duresco, water_paint, kunstnefarver, klee.

n

Number of colours needed. The largest palettes are parsons_1 and parsons_2 inspired by Thomas Parsons's colour samples from the book "A Tint Of Historical colours" (1934) and include 87 colours and 120 colours, respectively. du_barry (n=4), delft_ware (n=4), imari (n=7) palettes and several others are inspired by the book "The anatomy of colour" by Patrick Baty. Other palettes are inspired by art collection of Hermitage museum.

type

Either "continuous" or "discrete".

Value

A vector of colours

Examples

hermitage_palette("parsons_1")
#>  [1] "#602F6B" "#00cdcd" "#7A0646" "#FFD700" "#eae0c8" "#6a6c3b" "#93fe51"
#>  [8] "#003050" "#845860" "#eedd82" "#FFD04F" "#f06b5c" "#722F37" "#eaa6b2"
#> [15] "#B2F3AC" "#EDAC6E" "#bcd4e6" "#eac14f" "#92a1cf" "#f2efe6" "#29324f"
#> [22] "#ffff99" "#0b1307" "#fc8eac" "#a2b2bd" "#f6c79e" "#FAD6A5" "#EBB6CB"
#> [29] "#B3CEFF" "#2f2323" "#940000" "#d73b3e" "#5C7C99" "#800F08" "#2a2f23"
#> [36] "#816aa2" "#c14040" "#EADEB3" "#25b387" "#aa6c39" "#0BDA51" "#ffc87c"
#> [43] "#555D50" "#660c21" "#807059" "#0f52ba" "#ffe4c4" "#4fa985" "#D6E3B5"
#> [50] "#967bb6" "#3C3E3F" "#A83641" "#0d0d91" "#170e01" "#f4bfc7" "#993c0e"
#> [57] "#d9dde3" "#DBF1FD" "#CDD8D9" "#4997d0" "#eba832" "#b8c6b5" "#fbc030"
#> [64] "#585442" "#0047ab" "#808080" "#d0aca4" "#F5E6EA" "#726D5A" "#DFDC87"
#> [71] "#6d9a9b" "#AC7D0C" "#e25822" "#00A36C" "#f5f5dc" "#c54b8c" "#62483C"
#> [78] "#697D58" "#21421e" "#3B6AA0" "#C0C0C0" "#273e2d" "#8F00FF" "#3A5F0B"
#> [85] "#1034a6" "#99adbe" "#e25822"
#> attr(,"class")
#> [1] "palette"
#> attr(,"name")
#> [1] "parsons_1"